refatorando sistema e implementando autenticação com JWT

+++ src/main/java/com/jubasbackend/controller/request/SpecialtyRequest.java
@@ -3,13 +3,18 @@ package com.jubasbackend.controller.request;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import io.swagger.v3.oas.annotations.media.Schema;
 
+import java.math.BigDecimal;
 import java.time.LocalTime;
 
 public record SpecialtyRequest(
         String name,
+
         @Schema(type = "string", pattern = "HH:mm")
         @JsonFormat(pattern = "HH:mm")
         LocalTime timeDuration,
-        Float price,
-        Short categoryId) {
+
+        BigDecimal price,
+
+        Short categoryId
+) {
 }
